requirements.txt
alembic==1.13.1
bcrypt==4.1.3
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==1.4.32
typing_extensions==4.12.2
Werkzeug==2.0.3
